# fix(core): only find nullifiers in valid epochs