@vivianjeng 2024-08-29
| benchmark (72959 epochs) | old | new | 
|--|--|--|
| first time | 48855 ms|  **29004 ms** |
| saved data |19953 ms |  **264 ms** |

> there are 35 valid epochs in 72959 epochs